Remove functionality other than cache path
