+++ index.js
@@ -10,10 +10,10 @@ const execFile = require('child_process').execFile;
 const env = process.env;
 const user = env.LOGNAME || env.USER || env.LNAME || env.USERNAME || '';
 const exclusions = ['--help'];
-const envPaths = require('./env-paths.js');
+const cachePaths = require('./cache-paths.js');
 
 const configfile = '.v8flags.'+process.versions.v8+'.'+crypto.createHash('md5').update(user).digest('hex')+'.json';
-const configPath = envPaths('js-v8flags').cache;
+const configPath = cachePaths('js-v8flags');
 
 const failureMessage = [
   'Unable to cache a config file for v8flags to a your home directory',
