+++ test.js
@@ -85,7 +85,7 @@ describe('v8flags', function () {
     env.HOME = path.join(__dirname, 'does-not-exist');
     // Clear require cached modules so the modified environment variable HOME is used
     delete require.cache[require.resolve('./')]
-    delete require.cache[require.resolve('./env-paths.js')]
+    delete require.cache[require.resolve('./cache-paths.js')]
     var v8flags = require('./');
     v8flags.configPath = env.HOME;
     var configfile = path.resolve(os.tmpdir(), v8flags.configfile);
